todoList/requirements.txt
alabaster==0.7.12
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-rest-swagger==2.2.0
djangorestframework==3.12.4
docutils==0.18
drf-yasg==1.20.0
idna==3.3
imagesize==1.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.2
pycodestyle==2.8.0
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
snowballstemmer==2.1.0
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
sqlparse==0.4.2
swagger-spec-validator==2.7.4
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
